Fix tests for floats now returned by the parser

+++ test/plugin/test_in_gelf.rb
@@ -63,7 +63,7 @@ class GelfInputTest < Test::Unit::TestCase
       emits.each_index { |i|
         puts emits[i].to_s
         assert_equal 'gelf', emits[i][0]
-        assert_equal tests[i][:timestamp].to_i, emits[i][1] unless tests[i][:timestamp].nil?
+        assert_equal tests[i][:timestamp].to_f, emits[i][1] unless tests[i][:timestamp].nil?
         assert_equal tests[i][:short_message], emits[i][2]['short_message']
         assert_equal tests[i][:full_message], emits[i][2]['full_message']
       }
